Make outputs match proxmark3
The old code was almost, but not quite, doing simple addition of large (48-bit) numbers.  However, the carry bit was not properly handled.  Specifically, when the temporary 16-bit `carry_sum` would have low eight bits of `0xff` and `carry_sum` from prior addition was non-zero.

Frankly, I'm not sure if this is a bug, or intentionally slightly different from just doing addition and checking the carry flag?

It just looked like it was 99.5% implementing addition of multi-byte values.  Then, the proxmark code actually ***did*** just use multi-byte addition.   It could be a bug in this code (and fixed with this change), or maybe the KDF intentionally did this, and the author of the proxmark version of the code didn't notice this edge case, and thought the KDF was properly implementing addition?